+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/V1NetworkAPIGroupClient.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.kubernetes.client;
+
+import io.fabric8.kubernetes.api.model.networking.v1.DoneableIngress;
+import io.fabric8.kubernetes.api.model.networking.v1.DoneableIngressClass;
+import io.fabric8.kubernetes.api.model.networking.v1.DoneableNetworkPolicy;
+import io.fabric8.kubernetes.api.model.networking.v1.Ingress;
+import io.fabric8.kubernetes.api.model.networking.v1.IngressClass;
+import io.fabric8.kubernetes.api.model.networking.v1.IngressClassList;
+import io.fabric8.kubernetes.api.model.networking.v1.IngressList;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicy;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicyList;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.kubernetes.client.dsl.internal.networking.v1.IngressClassOperationsImpl;
+import io.fabric8.kubernetes.client.dsl.internal.networking.v1.IngressOperationsImpl;
+import io.fabric8.kubernetes.client.dsl.internal.networking.v1.NetworkPolicyOperationsImpl;
+import okhttp3.OkHttpClient;
+
+public class V1NetworkAPIGroupClient extends BaseClient implements V1NetworkAPIGroupDSL {
+  public V1NetworkAPIGroupClient() {
+    super();
+  }
+
+  public V1NetworkAPIGroupClient(OkHttpClient httpClient, final Config config) {
+    super(httpClient, config);
+  }
+
+  @Override
+  public MixedOperation<NetworkPolicy, NetworkPolicyList, DoneableNetworkPolicy, Resource<NetworkPolicy, DoneableNetworkPolicy>> networkPolicies() {
+    return new NetworkPolicyOperationsImpl(httpClient, getConfiguration());
+  }
+
+  @Override
+  public MixedOperation<Ingress, IngressList, DoneableIngress, Resource<Ingress, DoneableIngress>> ingresses() {
+    return new IngressOperationsImpl(httpClient, getConfiguration());
+  }
+
+  @Override
+  public MixedOperation<IngressClass, IngressClassList, DoneableIngressClass, Resource<IngressClass, DoneableIngressClass>> ingressClasses() {
+    return new IngressClassOperationsImpl(httpClient, getConfiguration());
+  }
+}