+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/NetworkAPIGroupDSL.java
@@ -15,15 +15,19 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
-import io.fabric8.kubernetes.api.model.networking.DoneableNetworkPolicy;
-import io.fabric8.kubernetes.api.model.networking.NetworkPolicy;
-import io.fabric8.kubernetes.api.model.networking.NetworkPolicyList;
+import io.fabric8.kubernetes.api.model.networking.v1.DoneableNetworkPolicy;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicy;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicyList;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.DoneableIngress;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.Ingress;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.IngressList;
 import io.fabric8.kubernetes.client.Client;
+import io.fabric8.kubernetes.client.V1NetworkAPIGroupDSL;
+import io.fabric8.kubernetes.client.V1beta1NetworkAPIGroupDSL;
 
 public interface NetworkAPIGroupDSL extends Client {
+  V1NetworkAPIGroupDSL v1();
+  V1beta1NetworkAPIGroupDSL v1beta1();
 
   MixedOperation<NetworkPolicy, NetworkPolicyList, DoneableNetworkPolicy, Resource<NetworkPolicy, DoneableNetworkPolicy>> networkPolicies();
   MixedOperation<Ingress, IngressList, DoneableIngress, Resource<Ingress, DoneableIngress>> ingress();