+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/NetworkPolicyExample.java
@@ -15,52 +15,53 @@
  */
 package io.fabric8.kubernetes.examples;
 
-import io.fabric8.kubernetes.api.model.networking.NetworkPolicy;
-import io.fabric8.kubernetes.api.model.networking.NetworkPolicyBuilder;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicy;
+import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicyBuilder;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.rmi.CORBA.Util;
-import java.util.HashMap;
+import java.util.Collections;
 
 /**
  * Created by hshinde on 1/6/18.
  */
 public class NetworkPolicyExample {
-  private static final Logger logger = LoggerFactory.getLogger(PodLogExample.class);
+  private static final Logger logger = LoggerFactory.getLogger(NetworkPolicyExample.class);
 
   public static void main(String[] args) {
     String namespace = "myproject";
 
-    KubernetesClient client = new DefaultKubernetesClient();
-
-    NetworkPolicy networkPolicy = new NetworkPolicyBuilder()
-                                        .withNewMetadata()
-                                          .withName("test-network-policy")
-                                        .endMetadata()
-                                        .withNewSpec()
-                                        .withNewPodSelector()
-                                          .withMatchLabels( new HashMap<String, String>() {{
-                                            put("role","db");
-                                          }})
-                                        .endPodSelector()
-                                        .endSpec()
-                                  .build();
+    try (KubernetesClient client = new DefaultKubernetesClient()) {
+      NetworkPolicy networkPolicy = new NetworkPolicyBuilder()
+        .withNewMetadata()
+        .withName("test-network-policy")
+        .endMetadata()
+        .withNewSpec()
+        .withNewPodSelector()
+        .withMatchLabels(Collections.singletonMap("role","db"))
+        .endPodSelector()
+        .endSpec()
+        .build();
 
 
-    // create polict using NetworkPolicyBuilder object
-    networkPolicy = client.network()
-      .networkPolicies()
-      .create(networkPolicy);
+      // create policy using NetworkPolicyBuilder object
+      networkPolicy = client.network()
+        .v1()
+        .networkPolicies()
+        .create(networkPolicy);
+      logger.info("NetworkPolicy {} created via builders", networkPolicy.getMetadata().getName());
 
-    // crate policy using YAML reosurce
-    networkPolicy = client.network()
-                                    .networkPolicies()
-                                    .inNamespace(namespace)
-                                    .load(NetworkPolicyExample.class.getResourceAsStream("/network-policy.yml"))
-                                    .createOrReplace();
+      // crate policy using YAML resource
+      networkPolicy = client.network()
+        .v1()
+        .networkPolicies()
+        .inNamespace(namespace)
+        .load(NetworkPolicyExample.class.getResourceAsStream("/network-policy.yml"))
+        .createOrReplace();
+      logger.info("NetworkPolicy {} created via YAML manifest", networkPolicy.getMetadata().getName());
 
+    }
   }
 }